jbuf: add delay wish size for start
At start of the call the jbuf_get returns silence until a specified buffer wish
size is reached. This can be used if a high jitter is expected and if an
improvement is possible. When later the network jitter improves the delay will
be reduced until the minimum value is reached.